moved files out of wds-website folder
