+++ app/favicon.ico
