--- wds-website/cached/newProjects.json
@@ -1,13 +1,0 @@
-[{
-    "name":"Dawnshipper",
-    "description":"In the enchanting realm of Sinia, where magic weaves through the air like a mystical tapestry, our protagonist, Devvyn Dawnshipper, emerges from the shadows of disgrace to weave his own destiny. Born into the shattered remnants of the once-proud Dawnshipper family, Devvyn carries the weight of public shame and personal vendettas on his broad shoulders. The haunting whispers of his family's disgrace follow him like spectral echoes through the hallowed halls of the Sinian Academy of Magic and Scientific Development.",
-    "image":null,
-    "type":"Novel"
-},
-{
-    "name":"Codename: Magic FPS",
-    "description":"Magic FPS is an arcade MOBA where players are free to explore/dictate their style of play with the provision of a multitude of artifacts with varying effects and abilities with varying number of paths.",
-    "image":null,
-    "type":"Game"
-}
-]
