--- wds-website/components/SplashLogo.tsx
@@ -1,73 +1,0 @@
-"use client";
-import {motion} from "framer-motion"
-import {  useState } from "react";
-export const Logo =()=>{
-    const [hasFdedIn,setHasFadedIn]=useState(false)
-    setTimeout(() => {
-        setHasFadedIn(true);
-    }, 2100);
-    return (
-     <motion.div  className=" absolute  left-[calc(50%+30px)] top-1/3 outline -translate-x-1/2 w-[400px] h-[350px] "   // Duration of the fade-in
-     >
-         <motion.svg
-         className="object-cover"
-         xmlns="http://www.w3.org/2000/svg"
-         version="1.1"
-         viewBox="0 0 4000 4000"
-         preserveAspectRatio="xMidYMid meet" // Adjusts scaling
-         style={{
-             shapeRendering: "geometricPrecision",
-             textRendering: "geometricPrecision",
-             fillRule: "evenodd",
-             clipRule: "evenodd",
-         }}
-         xmlnsXlink="http://www.w3.org/1999/xlink"
-    >
-        <defs>
-        <mask id="mask-gradient">
-          <motion.rect
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            fill="url(#grad1)"
-            initial={{ y: "100%" }}
-            animate={{ y: "45%" }}
-            transition={{ duration: 2, ease: "easeInOut" }}
-          />
-          <linearGradient id="grad1" x1="0%" y1="100%" x2="0%" y2="0%">
-            <stop offset="0%" style={{ stopColor: "#fefefe", stopOpacity: 0 }} />
-            <stop offset="100%" style={{ stopColor: "#fefefe", stopOpacity: 1 }} />
-          </linearGradient>
-        </mask>
-      </defs>
-      <g><motion.path initial={{opacity:0}} transition={hasFdedIn&&{duration:1,delay:0.2,ease:"easeOut",repeat:Infinity}} animate={hasFdedIn&&{opacity:[0.907],y:-300}} fill="#fefffe" d="M 1107.5,1655.5 C 1147.48,1695.15 1187.48,1734.81 1227.5,1774.5C 1228.83,1781.83 1228.83,1789.17 1227.5,1796.5C 1187.51,1757.01 1147.68,1717.35 1108,1677.5C 1107.5,1670.17 1107.33,1662.84 1107.5,1655.5 Z"/></g>
-            <g ><motion.path initial={{opacity:0,y:0}} transition={hasFdedIn&&{duration:1,delay:0.25,ease:"easeOut",repeat:Infinity}} animate={hasFdedIn&& {opacity:[0.966],y:-300}} fill="#fefffe" d="M 1107.5,1755.5 C 1147.8,1795.3 1187.97,1835.3 1228,1875.5C 1228.83,1895.84 1228.67,1916.17 1227.5,1936.5C 1187.15,1896.99 1147.32,1856.99 1108,1816.5C 1107.5,1796.17 1107.33,1775.84 1107.5,1755.5 Z"/></g>
-      {/* w below */}
-      <g mask="url(#mask-gradient)">
-        <motion.path
-          style={{ opacity: 0.994 }} fill="#fefffe" d="M 647.5,1875.5 C 687.5,1875.5 727.5,1875.5 767.5,1875.5C 767.333,1934.83 767.5,1994.17 768,2053.5C 824.04,1997.13 880.373,1941.13 937,1885.5C 993.654,1941.49 1050.32,1997.49 1107,2053.5C 1107.33,2027.5 1107.67,2001.5 1108,1975.5C 1148,2015.5 1188,2055.5 1228,2095.5C 1228.83,2175.5 1228.67,2255.5 1227.5,2335.5C 1130.82,2238.99 1033.99,2142.65 937,2046.5C 840.667,2142.83 744.333,2239.17 648,2335.5C 647.5,2182.17 647.333,2028.83 647.5,1875.5 Z"/></g>
-    
-    <g><path style={{opacity:0.906}} fill="#fefffe" d="M 1262.5,1875.5 C 1269.51,1875.33 1276.51,1875.5 1283.5,1876C 1323.35,1915.68 1363.01,1955.51 1402.5,1995.5C 1395.52,1996.66 1388.52,1996.83 1381.5,1996C 1341.25,1956.25 1301.58,1916.08 1262.5,1875.5 Z"/></g>
-    <g><motion.path  initial={{opacity:0,y:0}} transition={hasFdedIn&&{duration:1,delay:0.3,ease:"easeOut",repeat:Infinity}} animate={hasFdedIn&& {opacity:[0.969],y:-300}} fill="#fefffe" d="M 1107.5,1875.5 C 1147.8,1915.3 1187.97,1955.3 1228,1995.5C 1228.83,2022.51 1228.67,2049.51 1227.5,2076.5C 1187.51,2037.01 1147.68,1997.35 1108,1957.5C 1107.5,1930.17 1107.33,1902.84 1107.5,1875.5 Z"/></g>
-    <g><path style={{opacity:0.963}} fill="#fefffe" d="M 1361.5,1876.5 C 1382,1875.34 1402.66,1875.17 1423.5,1876C 1463.35,1915.68 1503.01,1955.51 1542.5,1995.5C 1522.17,1996.67 1501.84,1996.83 1481.5,1996C 1441.63,1955.97 1401.63,1916.13 1361.5,1876.5 Z"/></g>
-    {/* d below */}
-    <g mask="url(#mask-gradient)">
-        <motion.path
-          style={{ opacity: 0.994 }}
-          fill="#fefefe"
-          d="M 1482.5,1875.5 C 1509.5,1875.33 1536.5,1875.5 1563.5,1876C 1603.17,1915.67 1642.83,1955.33 1682.5,1995C 1688.5,1995.67 1694.5,1995.67 1700.5,1995C 1660.83,1955.33 1621.17,1915.67 1581.5,1876C 1648.5,1875.5 1715.5,1875.33 1782.5,1875.5C 1782.67,1982.83 1782.5,2090.17 1782,2197.5C 1756.17,2223.33 1730.33,2249.17 1704.5,2275C 1557.17,2275.42 1409.83,2275.92 1262.5,2276.5C 1262.5,2189.5 1262.5,2102.5 1262.5,2015.5C 1302.5,2015.5 1342.5,2015.5 1382.5,2015.5C 1382.5,2062.17 1382.5,2108.83 1382.5,2155.5C 1475.83,2155.5 1569.17,2155.5 1662.5,2155.5C 1662.5,2102.5 1662.5,2049.5 1662.5,1996.5C 1642.16,1996.67 1621.83,1996.5 1601.5,1996C 1561.25,1956.25 1521.58,1916.08 1482.5,1875.5 Z"
-        />
-      </g>
-    {/* s below */}
-    <g mask="url(#mask-gradient)">
-        <motion.path
-           style={{opacity:0.995}} fill="#fefffe" d="M 1817.5,1875.5 C 1844.5,1875.33 1871.5,1875.5 1898.5,1876C 1938.5,1916 1978.5,1956 2018.5,1996C 1991.5,1996.5 1964.5,1996.67 1937.5,1996.5C 1937.5,2002.83 1937.5,2009.17 1937.5,2015.5C 2070.83,2015.5 2204.17,2015.5 2337.5,2015.5C 2337.5,2102.17 2337.5,2188.83 2337.5,2275.5C 2164.17,2275.5 1990.83,2275.5 1817.5,2275.5C 1836.48,2235.38 1856.15,2195.54 1876.5,2156C 1990.17,2155.5 2103.83,2155.33 2217.5,2155.5C 2217.5,2149.17 2217.5,2142.83 2217.5,2136.5C 2084.17,2136.5 1950.83,2136.5 1817.5,2136.5C 1817.5,2049.5 1817.5,1962.5 1817.5,1875.5 Z"/></g>
-    <g><path style={{opacity:0.971}} fill="#fefffe" d="M 1916.5,1876.5 C 1943.66,1875.34 1971,1875.17 1998.5,1876C 2038.35,1915.68 2078.01,1955.51 2117.5,1995.5C 2090.51,1996.67 2063.51,1996.83 2036.5,1996C 1996.63,1955.97 1956.63,1916.13 1916.5,1876.5 Z"/></g>
-    <g><path style={{opacity:0.963}} fill="#fefffe" d="M 2057.5,1875.5 C 2077.84,1875.33 2098.17,1875.5 2118.5,1876C 2158.5,1916 2198.5,1956 2238.5,1996C 2217.83,1996.67 2197.17,1996.67 2176.5,1996C 2136.25,1956.25 2096.58,1916.08 2057.5,1875.5 Z"/></g>
-    <g><path style={{opacity:0.905}} fill="#fefffe" d="M 2196.5,1876.5 C 2203.65,1875.34 2210.99,1875.18 2218.5,1876C 2258.35,1915.68 2298.01,1955.51 2337.5,1995.5C 2330.52,1996.66 2323.52,1996.83 2316.5,1996C 2276.63,1955.97 2236.63,1916.13 2196.5,1876.5 Z"/></g>
-    
-          </motion.svg>
-     </motion.div>
-    )
-  }
